+++ src/basic/domains/cart/types/entities.ts
@@ -1,0 +1,11 @@
+import type { Product } from "../../../../types";
+
+export interface CartItem {
+  product: Product;
+  quantity: number;
+}
+
+export interface CartTotals {
+  totalBeforeDiscount: number;
+  totalAfterDiscount: number;
+}
